+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.9.1
 - Added trainer loss command, command that is run when the player loses to the trainer. Syntax: /trainers \<name\> setlosscommand \<command\>
 - Made trainer data be saved on update, so data is not lost on a server crash.
+- Added shiny selection
 
 # 0.9.0
 - Added nature selection